+++ src/main/java/mainDuke/StageStuff/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package mainDuke.StageStuff;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -27,7 +27,7 @@ public class DialogBox extends HBox {
 
     private DialogBox(String text, Image img) {
         try {
-            FXMLLoader fxmlLoader = new FXMLLoader(duke.MainWindow.class.getResource("/view/DialogBox.fxml"));
+            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
             fxmlLoader.setController(this);
             fxmlLoader.setRoot(this);
             fxmlLoader.load();
@@ -37,6 +37,7 @@ public class DialogBox extends HBox {
 
         dialog.setText(text);
         displayPicture.setImage(img);
+        assert displayPicture != null : "displayPicture should not be null";
     }
 
     /**
