+++ src/main/java/mainDuke/StageStuff/Main.java
@@ -1,7 +1,8 @@
-package duke;
+package mainDuke.StageStuff;
 
 import java.io.IOException;
 
+import mainDuke.Duke;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
